pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = ""
authors = ["Stefan Trawicki <s.trawicki@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic = "^2.5.2"
jinja2 = "^3.1.2"
pytailwindcss = "^0.2.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
